# openapi.yml
openapi: 3.1.0
info:
  title: NCANode API
  description: |-
    Server-side подпись и верификация для ЭЦП НУЦ РК (gosrk).

    Интегрируется с НУЦ-инфраструктурой: pki.gov.kz / test.pki.gov.kz
    (OCSP/TSP/CRL). Поддерживает GOST 34.10-2015 (актуальный стандарт),
    форматы CMS/CAdES-T, XMLDSIG, PAdES, WS-Security, JWT.

    Каждая операция доступна в одиночном (`/<service>/<op>`)
    и batch (`/<service>/<op>/batch`) вариантах.
  license:
    name: MIT
    url: https://github.com/malikzh/NCANode/blob/master/LICENSE
  version: v4
servers:
- url: ""
  description: current server
...
paths:
...
  /xml/verify/batch:
    post:
      tags:
      - XML
      operationId: verifyBatch
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/XmlVerifyBatchRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/XmlVerifyBatchResponse"
...
  /xml/sign/batch:
    post:
      tags:
      - XML
      operationId: signBatch
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/XmlSignBatchRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/XmlSignBatchResponse"
  /xades/verify:
    post:
      tags:
      - XAdES
      operationId: verify_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/XadesVerifyRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/XadesVerificationResponse"
  /xades/verify/batch:
    post:
      tags:
      - XAdES
      operationId: verifyBatch_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/XadesVerifyBatchRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/XadesVerificationBatchResponse"
  /xades/sign:
    post:
      tags:
      - XAdES
      operationId: sign_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/XadesSignRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/XadesResponse"
  /xades/sign/batch:
    post:
      tags:
      - XAdES
      operationId: signBatch_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/XadesSignBatchRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/XadesSignBatchResponse"
  /xades/extend:
    post:
      tags:
      - XAdES
      operationId: extend
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/XadesExtendRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/XadesResponse"
...
  /x509/verify/batch:
    post:
      tags:
      - X509
      operationId: verifyBatch_2
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SbaVerifyBatchRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SbaVerifyBatchResponse"
...
  /wsse/verify/batch:
    post:
      tags:
      - WSSE
      operationId: verifyBatch_3
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WsseVerifyBatchRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/WsseVerifyBatchResponse"
...
  /wsse/sign/batch:
    post:
      tags:
      - WSSE
      operationId: signBatch_2
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WsseSignBatchRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/WsseSignBatchResponse"
...
  /pdf/verify/batch:
    post:
      tags:
      - PDF
      operationId: verifyBatch_4
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PdfVerifyBatchRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PdfVerifyBatchResponse"
...
  /pdf/sign/batch:
    post:
      tags:
      - PDF
      operationId: signBatch_3
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PdfSignBatchRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PdfSignBatchResponse"
  /pades/verify:
    post:
      tags:
      - PAdES
      operationId: verify_5
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PadesVerifyRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PadesVerificationResponse"
  /pades/verify/batch:
    post:
      tags:
      - PAdES
      operationId: verifyBatch_5
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PadesVerifyBatchRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PadesVerificationBatchResponse"
  /pades/sign:
    post:
      tags:
      - PAdES
      operationId: sign_4
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PadesSignRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PadesResponse"
  /pades/sign/batch:
    post:
      tags:
      - PAdES
      operationId: signBatch_4
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PadesSignBatchRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PadesSignBatchResponse"
  /pades/extend:
    post:
      tags:
      - PAdES
      operationId: extend_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PadesExtendRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PadesResponse"
...
  /cms/verify/batch:
    post:
      tags:
      - CMS
      operationId: verifyBatch_6
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CmsVerifyBatchRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CmsVerificationBatchResponse"
  /cms/sign:
    post:
      tags:
      - CMS
      operationId: sign_5
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CmsCreateRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CmsResponse"
    patch:
      tags:
      - CMS
      operationId: signPatch
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CmsCreateRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CmsResponse"
  /cms/sign/batch:
    post:
      tags:
      - CMS
      operationId: signBatch_5
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CmsCreateBatchRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CmsBatchResponse"
...
  /cades/verify:
    post:
      tags:
      - CAdES
      operationId: verify_7
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CadesVerifyRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CadesVerificationResponse"
  /cades/verify/batch:
    post:
      tags:
      - CAdES
      operationId: verifyBatch_7
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CadesVerifyBatchRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CadesVerificationBatchResponse"
  /cades/sign:
    post:
      tags:
      - CAdES
      operationId: sign_6
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CadesSignRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CadesResponse"
    patch:
      tags:
      - CAdES
      operationId: coSign
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CadesSignRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CadesResponse"
  /cades/sign/batch:
    post:
      tags:
      - CAdES
      operationId: signBatch_6
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CadesSignBatchRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CadesSignBatchResponse"
  /cades/extend:
    post:
      tags:
      - CAdES
      operationId: extend_2
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CadesExtendRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CadesResponse"
...
components:
  schemas:
...
    XmlVerifyBatchRequest:
      type: object
      properties:
        revocationCheck:
          type: array
          items:
            type: string
            enum:
            - OCSP
            - CRL
          uniqueItems: true
        xmls:
          type: array
          items:
            type: string
          minItems: 1
      required:
      - xmls
    XmlVerifyBatchResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
        message:
          type: string
        results:
          type: array
          items:
            $ref: "#/components/schemas/VerificationResponse"
...
    XmlSignBatchRequest:
      type: object
      properties:
        xmls:
          type: array
          items:
            type: string
          minItems: 1
        signers:
          type: array
          items:
            $ref: "#/components/schemas/SignerRequest"
          minItems: 1
        trimXml:
          type: boolean
        clearSignatures:
          type: boolean
      required:
      - signers
      - xmls
...
    XmlSignBatchResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
        message:
          type: string
        results:
          type: array
          items:
            $ref: "#/components/schemas/Item"
    XadesVerifyRequest:
      type: object
      properties:
        revocationCheck:
          type: array
          items:
            type: string
            enum:
            - OCSP
            - CRL
          uniqueItems: true
        xml:
          type: string
          minLength: 1
      required:
      - xml
...
    XadesVerificationResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
        message:
          type: string
        valid:
          type: boolean
        level:
          type: string
          enum:
          - B
          - T
          - LT
          - LTA
        verifiedLevel:
          type: string
          enum:
          - B
          - T
          - LT
          - LTA
        signatures:
          type: array
          items:
            $ref: "#/components/schemas/XadesSignatureInfo"
    XadesVerifyBatchRequest:
      type: object
      properties:
        revocationCheck:
          type: array
          items:
            type: string
            enum:
            - OCSP
            - CRL
          uniqueItems: true
        xmls:
          type: array
          items:
            type: string
          minItems: 1
      required:
      - xmls
    XadesVerificationBatchResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
        message:
          type: string
        results:
          type: array
          items:
            $ref: "#/components/schemas/XadesVerificationResponse"
    XadesSignRequest:
      type: object
      properties:
        xml:
          type: string
          minLength: 1
        signers:
          type: array
          items:
            $ref: "#/components/schemas/SignerRequest"
          minItems: 1
        level:
          type: string
          enum:
          - B
          - T
          - LT
          - LTA
        packaging:
          type: string
          enum:
          - ENVELOPED
          - ENVELOPING
          - DETACHED
        tsaPolicy:
          type: string
          enum:
          - TSA_GOST2015_POLICY
          - TSA_GOST_POLICY
          - TSA_GOSTGT_POLICY
      required:
      - signers
      - xml
    XadesResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
        message:
          type: string
        xml:
          type: string
        level:
          type: string
          enum:
          - B
          - T
          - LT
          - LTA
    XadesSignBatchRequest:
      type: object
      properties:
        xmls:
          type: array
          items:
            type: string
          minItems: 1
        signers:
          type: array
          items:
            $ref: "#/components/schemas/SignerRequest"
          minItems: 1
        level:
          type: string
          enum:
          - B
          - T
          - LT
          - LTA
        packaging:
          type: string
          enum:
          - ENVELOPED
          - ENVELOPING
          - DETACHED
        tsaPolicy:
          type: string
          enum:
          - TSA_GOST2015_POLICY
          - TSA_GOST_POLICY
          - TSA_GOSTGT_POLICY
      required:
      - signers
      - xmls
    XadesSignBatchResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
        message:
          type: string
        results:
          type: array
          items:
            $ref: "#/components/schemas/Item"
    XadesExtendRequest:
      type: object
      properties:
        xml:
          type: string
          minLength: 1
        level:
          type: string
          enum:
          - B
          - T
          - LT
          - LTA
        tsaPolicy:
          type: string
          enum:
          - TSA_GOST2015_POLICY
          - TSA_GOST_POLICY
          - TSA_GOSTGT_POLICY
      required:
      - xml
...
    SbaVerifyBatchRequest:
      type: object
      properties:
        revocationCheck:
          type: array
          items:
            type: string
            enum:
            - OCSP
            - CRL
          uniqueItems: true
        items:
          type: array
          items:
            $ref: "#/components/schemas/Item"
          minItems: 1
      required:
      - items
    SbaVerifyBatchResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
        message:
          type: string
        results:
          type: array
          items:
            $ref: "#/components/schemas/VerificationResponse"
...
    WsseVerifyBatchRequest:
      type: object
      properties:
        revocationCheck:
          type: array
          items:
            type: string
            enum:
            - OCSP
            - CRL
          uniqueItems: true
        xmls:
          type: array
          items:
            type: string
          minItems: 1
      required:
      - xmls
    WsseVerifyBatchResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
        message:
          type: string
        results:
          type: array
          items:
            $ref: "#/components/schemas/VerificationResponse"
...
    WsseSignBatchRequest:
      type: object
      properties:
        xmls:
          type: array
          items:
            type: string
          minItems: 1
        key:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 1
        keyAlias:
          type: string
        trimXml:
          type: boolean
      required:
      - key
      - password
      - xmls
    WsseSignBatchResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
        message:
          type: string
        results:
          type: array
          items:
            $ref: "#/components/schemas/Item"
...
    PdfVerifyBatchRequest:
      type: object
      properties:
        revocationCheck:
          type: array
          items:
            type: string
            enum:
            - OCSP
            - CRL
          uniqueItems: true
        pdfs:
          type: array
          items:
            type: string
          minItems: 1
      required:
      - pdfs
    PdfVerifyBatchResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
        message:
          type: string
        results:
          type: array
          items:
            $ref: "#/components/schemas/PdfVerificationResponse"
...
    PdfSignBatchRequest:
      type: object
      properties:
        pdfs:
          type: array
          items:
            type: string
          minItems: 1
        signers:
          type: array
          items:
            $ref: "#/components/schemas/PdfSigner"
          minItems: 1
        tsaPolicy:
          type: string
          enum:
          - TSA_GOST2015_POLICY
          - TSA_GOST_POLICY
          - TSA_GOSTGT_POLICY
        withTsp:
          type: boolean
      required:
      - pdfs
      - signers
    PdfSignBatchResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
        message:
          type: string
        results:
          type: array
          items:
            $ref: "#/components/schemas/Item"
    PadesVerifyRequest:
      type: object
      properties:
        revocationCheck:
          type: array
          items:
            type: string
            enum:
            - OCSP
            - CRL
          uniqueItems: true
        pdf:
          type: string
          minLength: 1
      required:
      - pdf
...
    PadesVerificationResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
        message:
          type: string
        valid:
          type: boolean
        level:
          type: string
          enum:
          - B
          - T
          - LT
          - LTA
        verifiedLevel:
          type: string
          enum:
          - B
          - T
          - LT
          - LTA
        signatures:
          type: array
          items:
            $ref: "#/components/schemas/PadesSignatureInfo"
        documentTimestamps:
          type: integer
          format: int32
    PadesVerifyBatchRequest:
      type: object
      properties:
        revocationCheck:
          type: array
          items:
            type: string
            enum:
            - OCSP
            - CRL
          uniqueItems: true
        pdfs:
          type: array
          items:
            type: string
          minItems: 1
      required:
      - pdfs
    PadesVerificationBatchResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
        message:
          type: string
        results:
          type: array
          items:
            $ref: "#/components/schemas/PadesVerificationResponse"
    PadesSignRequest:
      type: object
      properties:
        pdf:
          type: string
          minLength: 1
        signers:
          type: array
          items:
            $ref: "#/components/schemas/PadesSigner"
          minItems: 1
        level:
          type: string
          enum:
          - B
          - T
          - LT
          - LTA
        tsaPolicy:
          type: string
          enum:
          - TSA_GOST2015_POLICY
          - TSA_GOST_POLICY
          - TSA_GOSTGT_POLICY
      required:
      - pdf
      - signers
...
    PadesResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
        message:
          type: string
        pdf:
          type: string
        level:
          type: string
          enum:
          - B
          - T
          - LT
          - LTA
    PadesSignBatchRequest:
      type: object
      properties:
        pdfs:
          type: array
          items:
            type: string
          minItems: 1
        signers:
          type: array
          items:
            $ref: "#/components/schemas/PadesSigner"
          minItems: 1
        level:
          type: string
          enum:
          - B
          - T
          - LT
          - LTA
        tsaPolicy:
          type: string
          enum:
          - TSA_GOST2015_POLICY
          - TSA_GOST_POLICY
          - TSA_GOSTGT_POLICY
      required:
      - pdfs
      - signers
    PadesSignBatchResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
        message:
          type: string
        results:
          type: array
          items:
            $ref: "#/components/schemas/Item"
    PadesExtendRequest:
      type: object
      properties:
        pdf:
          type: string
          minLength: 1
        level:
          type: string
          enum:
          - B
          - T
          - LT
          - LTA
        tsaPolicy:
          type: string
          enum:
          - TSA_GOST2015_POLICY
          - TSA_GOST_POLICY
          - TSA_GOSTGT_POLICY
      required:
      - pdf
...
    CmsVerifyBatchRequest:
      type: object
      properties:
        revocationCheck:
          type: array
          items:
            type: string
            enum:
            - OCSP
            - CRL
          uniqueItems: true
        items:
          type: array
          items:
            $ref: "#/components/schemas/Item"
          minItems: 1
      required:
      - items
    CmsVerificationBatchResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
        message:
          type: string
        results:
          type: array
          items:
            $ref: "#/components/schemas/CmsVerificationResponse"
    CmsCreateRequest:
      type: object
      properties:
        cms:
          type: string
        data:
          type: string
        signers:
          type: array
          items:
            $ref: "#/components/schemas/SignerRequest"
          minItems: 1
        tsaPolicy:
          type: string
          enum:
          - TSA_GOST2015_POLICY
          - TSA_GOST_POLICY
          - TSA_GOSTGT_POLICY
        detached:
          type: boolean
        withTsp:
          type: boolean
      required:
      - signers
    CmsResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
        message:
          type: string
        cms:
          type: string
    CmsCreateBatchRequest:
      type: object
      properties:
        data:
          type: array
          items:
            type: string
          minItems: 1
        signers:
          type: array
          items:
            $ref: "#/components/schemas/SignerRequest"
          minItems: 1
        tsaPolicy:
          type: string
          enum:
          - TSA_GOST2015_POLICY
          - TSA_GOST_POLICY
          - TSA_GOSTGT_POLICY
        detached:
          type: boolean
        withTsp:
          type: boolean
      required:
      - data
      - signers
    CmsBatchResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
        message:
          type: string
        results:
          type: array
          items:
            $ref: "#/components/schemas/Item"
...
    CadesVerifyRequest:
      type: object
      properties:
        revocationCheck:
          type: array
          items:
            type: string
            enum:
            - OCSP
            - CRL
          uniqueItems: true
        cms:
          type: string
          minLength: 1
        data:
          type: string
      required:
      - cms
...
    CadesVerificationResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
        message:
          type: string
        valid:
          type: boolean
        level:
          type: string
          enum:
          - B
          - T
          - LT
          - LTA
        verifiedLevel:
          type: string
          enum:
          - B
          - T
          - LT
          - LTA
        signers:
          type: array
          items:
            $ref: "#/components/schemas/CadesSignerInfo"
    CadesVerifyBatchRequest:
      type: object
      properties:
        revocationCheck:
          type: array
          items:
            type: string
            enum:
            - OCSP
            - CRL
          uniqueItems: true
        items:
          type: array
          items:
            $ref: "#/components/schemas/Item"
          minItems: 1
      required:
      - items
...